requirements.txt
absl-py==0.9.0
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
bcrypt==3.1.6
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.12
defer==1.0.6
distro==1.3.0
distro-info===0.21ubuntu4
duplicity==0.8.4
easydict==1.9
entrypoints==0.3
fasteners==0.12.0
Flask==1.1.2
future==0.16.0
gast==0.3.3
google-pasta==0.2.0
google-speech==1.1.0
grpcio==1.30.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
httplib2==0.11.3
idna==2.6
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.26
jsonpointer==2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.10.0
macaroonbakery==1.2.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.3.0
mongoengine==0.20.0
monotonic==1.5
netifaces==0.10.4
numpy==1.19.0
oauthlib==2.1.0
olefile==0.46
opencv-python==4.2.0.34
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.2.0
playsound==1.2.2
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.34.0
PyJWT==1.7.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==1.9.0+ubuntu1.4
python-dateutil==2.7.3
python-debian==0.1.36
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
pyzmq==19.0.2
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
scipy==1.5.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.18
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.5.1
torchfile==0.1.0
torchvision==0.6.1
tornado==6.0.4
tqdm==4.47.0
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.7
visdom==0.1.8.9
wadllib==1.3.3
web-cache==1.1.0
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
